wip on standalone

+++ contracts/interpreter/IInterpreter.sol
@@ -4,5 +4,9 @@ pragma solidity ^0.8.0;
 type SourceIndex is uint256;
 
 interface IInterpreter {
-    function eval(address expressionPointer, SourceIndex entrypoint, uint[][] memory context) external view returns (uint[] memory);
+    function eval(
+        address expressionPointer,
+        SourceIndex entrypoint,
+        uint256[][] memory context
+    ) external view returns (uint256[] memory);
 }
